fix warning

+++ src/pg9X/init-pg9X.py
@@ -143,7 +143,8 @@ logfile = os.path.join(pg_log, "install.log")
 
 if util.get_platform() == "Linux":
   if not os.path.isfile("/usr/lib64/libtinfo.so.5"):
-    os.system("sudo ln -s /usr/lib64/libtinfo.so.6 /usr/lib64/libtinfo.so.5")
+    if os.path.isfile("/usr/lib64/libtinfo.so.6"):
+      os.system("sudo ln -s /usr/lib64/libtinfo.so.6 /usr/lib64/libtinfo.so.5")
 
 ## INITDB #############################################
 print('\nInitializing Postgres DB with:')
